+++ src/LaMEM.jl
@@ -19,6 +19,38 @@ using .Run
 export run_lamem, run_lamem_save_grid, mpiexec
 export remove_popup_messages_mac, show_paths_LaMEM
 
+
+# ==
+using DocStringExtensions
+
+"""
+    help_info::String = get_doc(structure, field::Symbol) 
+This returns a string with the documentation for a parameter `field` that is within the `structure`. 
+Note that this structure must be a help structure of the current one.
+"""
+function get_doc(structure, field::Symbol) 
+    alldocs       =   Docs.meta(LaMEM.LaMEM_Model);
+    #alldocs       =   Docs.meta();
+    
+    var           =   eval(Meta.parse("Docs.@var($structure)"))
+    @show var
+    #=
+    fields_local  =   alldocs[var].docs[Union{}].data[:fields]
+    str = fields_local[field]
+
+    # Add comment to next line (if required)
+    str = replace(str, "\n" => "\n #")
+
+    # remove the # at the end of the string
+    if str[end]=='#'
+        str = str[1:end-1]
+    end
+=#
+    str="#"
+    return str
+end
+# ==
+
 # Functions that help running LaMEM directly from julia
 include("LaMEM_ModelGeneration/LaMEM_Model.jl")
 using .LaMEM_Model
